requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
async-timeout==5.0.1
asyncpg==0.31.0
exceptiongroup==1.3.1
fastapi==0.122.0
greenlet==3.2.4
idna==3.11
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
python-dotenv==1.2.1
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
